Move schema related files into schema directory
